refactor: the event handlers
